allow execution via `mix run`
Prior to this commit erl-dns would need to be run with the packaged
`run.sh` script, which would need to be updated with the ebin
locations for the locally installed Elixir. This commit converts the
`erldns.config.example` into a mix config that is read during
compilation and included in the app, which enables erl-dns to now be run
via `mix run --no-halt` or `iex -S mix run`.